requirements.txt
python-dotenv
google-cloud-storage
google-genai